requirements.txt
numpy==1.24.3
tensorflow==2.12.0
scikit-learn==1.2.2
pillow==9.5.0